ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
scikit-learn==1.5.0
numpy==1.26.4
httpx==0.27.0
reportlab==4.2.0
python-dotenv==1.0.1
